Add files via upload
This report aims to assist Southwark Council to address the continuously increasing bicycle thefts in London Borough of Southwark and promote cycling as a sustainable mode of transportation. 

To address the issue of bicycle theft, the locations for each recorded bicycle theft in London from December 2013 to November 2014 were provided as the basis for analysis. The report focuses on the hotspots, locations, trends, count and rates of bicycle theft, followed by a discussion of limitations and suggestions to provide practitioners with a reference for decision-making.